requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
boto3==1.37.7
botocore==1.37.7
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chess==1.11.2
click==8.1.8
distro==1.9.0
e2b==1.1.0
e2b-code-interpreter==1.0.5
fastapi==0.115.11
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.8.2
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mss==10.0.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.3
openai==1.65.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
playwright==1.50.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pyenchant==3.2.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
TextArena==0.5.8
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
yarl==1.18.3